task-86005 option_on_adding_new_files_with_Inv_or_catalog_will_also_be_attached_with_previous_order
Started work on the following points:

Add option with Digital attachments form as:

Attach With Existing Orders - (Yes/No)

so that when seller/admin attach new files/links and on selection of YES added attachment would also be added with ordered items. These newly added items will get display on order view page for all buyers.

Completed in admin section and tested the same. In process of adding in front end for seller.